Putting all interfaces to mutal ts file
