+++ sync-snack/app/types/types.tsx
@@ -1,3 +1,4 @@
+// All sort options we have for this app for leaderboard
 export enum SortOption {
     CoffeeCount = "coffeeCount",
     Rating = "rating",
